+++ config/routes.rb
@@ -1,5 +1,10 @@
 Rails.application.routes.draw do
   namespace :admin do
+    get "dashboard/index"
+    get "login" => "sessions#new"
+    post "login" => "sessions#create"
+    delete "logout" => "sessions#destroy"
+    get "dashboard" => "dashboard#index", as: :dashboard
     resources :tenants do
       resources :sites do
       end
@@ -16,10 +21,10 @@ Rails.application.routes.draw do
 
   get "/check_pnr", to: "pnumber#check_pnr", as: :check_pnr
   get "/check_phone", to: "pnumber#check_phone", as: :check_phone
-  
-  resource :session, only: [:create, :update]
+
+  resource :session, only: [ :create, :update ]
   get "/guest/s/default/", to: "sessions#new", as: :new_session, constraints: { format: "html" }
-  resolve("Session") { [:session] }
+  resolve("Session") { [ :session ] }
   get "otp", to: "sessions#otp", as: :otp
   get "success", to: "sessions#success", as: :success
   post "resend_otp", to: "sessions#resend", as: :resend_otp
@@ -34,6 +39,5 @@ Rails.application.routes.draw do
   # get "service-worker" => "rails/pwa#service_worker", as: :pwa_service_worker
 
   # Defines the root path route ("/")
-  root "home#index"
-
+  root "admin/dashboard#index"
 end